removed username in the steps
