+++ src/components/HowItWorks/StepList.tsx
@@ -5,10 +5,6 @@ export default function () {
   return (
     <OrderedList>
       <CardParagraph>
-        Verify that you have a profile on Farcaster by matching your username to
-        the wallet you have connected to Farcaster.
-      </CardParagraph>
-      <CardParagraph>
         Create your burner wallet with a Farcaster ZK badge.
       </CardParagraph>
       <CardParagraph>Create a cast that will be anonymous.</CardParagraph>
