+++ src/pages/CreateBurnerWallet.tsx
@@ -1,25 +1,11 @@
-import { AccentText, HeaderText, PostText } from 'components/ui/Text'
-import { Sizes } from 'models/CharInCircle'
-import { handleError } from '@big-whale-labs/frontend-utils'
+import { HeaderText, PostText } from 'components/ui/Text'
+import { Redirect } from 'wouter'
 import { useEffect, useState } from 'preact/hooks'
 import { useSnapshot } from 'valtio'
 import BurnerWalletCard from 'components/BurnerWalletCard'
 import BurnerWalletStore from 'stores/BurnerWalletStore'
-import Button from 'components/ui/Button'
-import CharInCircle from 'components/ui/CharInCircle'
 import HowItWorks from 'components/HowItWorks'
-import TextInput from 'components/ui/TextInput'
-import Tooltip from 'components/ui/Tooltip'
-import classnames, {
-  alignItems,
-  display,
-  flexDirection,
-  gap,
-  justifyContent,
-  textAlign,
-  width,
-} from 'classnames/tailwind'
-import getErrorMessage from 'helpers/getErrorMessage'
+import classnames, { display, flexDirection, gap } from 'classnames/tailwind'
 import walletStore from 'stores/WalletStore'
 
 const wrapper = classnames(
@@ -32,73 +18,22 @@ const topPart = classnames(
   flexDirection('flex-col'),
   gap('gap-y-4')
 )
-const bottomPart = classnames(
-  display('flex'),
-  flexDirection('flex-col'),
-  justifyContent('justify-start'),
-  gap('gap-y-6')
-)
-
-const hintWrapper = classnames(
-  display('inline-flex'),
-  alignItems('items-center'),
-  gap('gap-x-1.5')
-)
-const buttonWithStatus = classnames(
-  display('flex'),
-  flexDirection('flex-col', 'sm:flex-row'),
-  alignItems('items-center'),
-  textAlign('text-center', 'sm:text-left'),
-  gap('gap-y-4', 'sm:gap-x-4')
-)
-const buttonClass = classnames(display('block'), width('w-full', 'sm:w-64'))
-const defaultMessage =
-  'We could not match the username with the wallet you have connected. Please review them and try again.'
 
 export default function () {
   const { account } = useSnapshot(walletStore)
   const { privateKey } = useSnapshot(BurnerWalletStore)
-  const [username, setUsername] = useState('')
   const [error, setError] = useState('')
-  const [loading, setLoading] = useState(false)
   const [status, setStatus] = useState<string | undefined>()
-
-  const generate = async () => {
-    setError('')
-    if (!username.length) return setError('Username cannot be empty')
-    setLoading(true)
-
-    try {
-      if (!walletStore.account) await walletStore.connect(true)
-      if (!walletStore.account) return setError('Please connect the wallet')
-      await BurnerWalletStore.generateBurnerWallet(
-        username,
-        walletStore.account,
-        setStatus
-      )
-      walletStore.exit()
-    } catch (error) {
-      const errorMessage = getErrorMessage(error)
-      setError(typeof errorMessage === 'string' ? errorMessage : defaultMessage)
-      handleError(error)
-    } finally {
-      setStatus(undefined)
-      setLoading(false)
-    }
-  }
 
   useEffect(() => {
     setStatus('')
     setError('')
   }, [account])
 
-  const hintText =
-    'You can see what eth address you have connected to Farcaster in profile settings.'
-
   return (
     <div className={wrapper}>
-      <div className={topPart}>
-        {privateKey ? (
+      {privateKey ? (
+        <div className={topPart}>
           <>
             <HeaderText>Voilà — your burner wallet</HeaderText>
             <PostText>
@@ -108,65 +43,11 @@ export default function () {
             </PostText>
             <BurnerWalletCard privateKey={privateKey} />
           </>
-        ) : (
-          <>
-            <HeaderText>
-              Now let’s verify that you have a Farcaster profile
-            </HeaderText>
-            <PostText>
-              We’ll need to match your Farcaster username with the wallet you
-              connected to your Farcaster profile.{' '}
-              <AccentText color="text-accent">
-                <Tooltip position="bottom" text={hintText}>
-                  <span className={hintWrapper}>
-                    <CharInCircle size={Sizes.Small} char="?" />
-                  </span>
-                </Tooltip>
-              </AccentText>
-            </PostText>
-          </>
-        )}
-      </div>
-      {privateKey ? (
-        <HowItWorks />
-      ) : (
-        <div className={bottomPart}>
-          <TextInput
-            withAtSign
-            value={username}
-            errorMessage={error}
-            disabled={loading}
-            onChange={({ target }) =>
-              setUsername((target as HTMLInputElement).value)
-            }
-            onKeyDown={({ code }) => code === 'Enter' && generate()}
-          />
-          <PostText>
-            You’ll verify your username and profile by connecting the same
-            wallet you connected to Farcaster.
-          </PostText>
-          <div className={buttonWithStatus}>
-            <div className={buttonClass}>
-              <Button
-                center
-                fullWidth
-                loadingOverflow
-                type="primary"
-                loading={loading}
-                disabled={!username.length}
-                onClick={generate}
-              >
-                Connect & verify
-              </Button>
-            </div>
-            {status && (
-              <AccentText small color="text-tertiary">
-                {status}
-              </AccentText>
-            )}
-          </div>
         </div>
+      ) : (
+        <Redirect to={'/'} />
       )}
+      <HowItWorks />
     </div>
-  )
+  ) //
 }
